Make work view page.tsx a server component and catid param required and singular
